Move v2 blocks into separate files (#1222, #1225)
